Refactor nexus_registry tests
Currently all the tests are put into one file.
It's hard to read and browse tests related to
different types and formats of repository provider.
There are also some missing checks.

This change addresses this problem and
- split tests
- add some todos
- rename helper functions
- fix some indents
